Django/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-admin-honeypot==1.1.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-user-agents==0.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==2.10
isodate==0.6.0
lxml==4.6.3
mysqlclient==2.0.3
oauthlib==3.1.1
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
ua-parser==0.10.0
urllib3==1.26.5
user-agents==2.2.0
zeep==4.0.0
